Try to use the correct format for ActionController::RoutingError responses w:@lleirborras
We cannot have the format for these requests, as they are not correct
routes - so we have no information about them.

The only format information is present in the Accept: HTTP header,
already handled by Rails, and by the last part of the URL. We try
here to infer whether this 'filename extension' matches a known MIME
type, and if this is the case we set it in the request format.